Update address display logic
displayAddress is now called inside rendering of search results.
This is done so that all data with each place can be saved instead
of just saving the address. As a result, lat and lon data can be
accessed much easier if the user chooses one of these results, which
can be used to change the location and update the page with the
weather for that location.