Update forecast scroll logic
This changes how the scroll works. Instead of using useState,
now use useRef. This is so that we can introduce eventListeners
that allow for scrolling through the hourly forecast without
needing to be in the window. As a result, we cannot use useState
in these eventListeners, and must use references.